checkers/requirements.txt
beautifulsoup4
checklib
diff-match-patch
gevent
lorem
lxml
pillow
requests
selenium
websocket_client
